requirements.txt
asgiref
certifi
Django
djangorestframework
dlib
numpy
opencv-python
Pillow
pytz
sqlparse
